- Multiple runtime fixes.  - Moved some fingerprint procs to detective_work.dm.  - Removed some unnecessary del() usage in the detective code.
git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5800 316c924e-a436-60f5-8080-3fe189b3f50e

+++ code/game/objects/effects/effect_system.dm
@@ -210,7 +210,8 @@ steam.start() -- spawns the effect
 					sleep(5)
 					step(sparks,direction)
 				spawn(20)
-					sparks.delete()
+					if(sparks) // Might be deleted already
+						sparks.delete()
 					src.total_sparks--
